deactivate quantized tests that use other backends
